refactor: 상태 데어티가 변하는 객체(또는 클래스)는 Model 폴더 내에서 관리
BridegGame 클래스로 부터 만들어진 인스턴스는 게임이 진행하면서 클래스 필드가 수시로 바뀐다.
