+++ docs/README.md
@@ -243,6 +243,7 @@ const number = generateRandomNumber();
   - 게임 진행을 담당하는 BridgeGameManager 클래스를 만들어 분리한다.
   - BridgeGameManager 클래스는 Model과 View를 연결하는 Controller의 역할을 한다.
   - 특정 기능을 지닌 객체 또는 로직은 libs 폴더 내에서 관리한다.
+  - 상태 데이터가 변하는 객체(또는 클래스)는 Model 폴더 내에서 관리한다.
 - [ ] refactor: 변수명, 함수명 수정하기
 - [ ] refactor: 매직 넘버 상수화하기
 - [ ] refactor: 함수(또는 메서드)가 한 가지의 일만 하도록 하기
